fix(isr,delay): address qodo task registration and overflow issues
- Add task handle verification after AD7609 task creation
  Ensures handle is properly populated, logs error if NULL
  Prevents silent failure of interrupt notifications

- Prevent delay arithmetic overflow in timer calculations
  Use 64-bit intermediate arithmetic in AD7609_Delay_ms and DAC7718_Delay_us
  Prevents overflow when large delay values * frequency exceed uint32_t
  Example: 1000ms * 100MHz would overflow without 64-bit cast

+++ firmware/src/HAL/ADC/AD7609.c
@@ -27,10 +27,11 @@
 #include "FreeRTOS.h"
 #include "task.h"
 
-// Simple delay function using core timer (wrap-around safe)
+// Simple delay function using core timer (wrap-around safe, overflow-safe)
 static void AD7609_Delay_ms(uint32_t milliseconds) {
     uint32_t startCount = CORETIMER_CounterGet();
-    uint32_t ticks = (milliseconds * (CORETIMER_FrequencyGet() / 1000U));
+    // Use 64-bit arithmetic to prevent overflow for large millisecond values
+    uint32_t ticks = (uint32_t)(((uint64_t)milliseconds * CORETIMER_FrequencyGet()) / 1000U);
 
     // Use modular arithmetic to handle 32-bit timer wrap-around correctly
     // Subtraction of unsigned values wraps correctly
